+++ src/blacksound/sampleprovider/pinknoisegenerator.h
@@ -33,7 +33,7 @@ namespace BlackSound
             CPinkNoiseGenerator(QObject *parent = nullptr) : ISampleProvider(parent) {}
 
             //! Read samples
-            virtual int readSamples(QVector<qint16> &samples, qint64 count) override;
+            virtual int readSamples(QVector<float> &samples, qint64 count) override;
 
             //! Gain
             void setGain(double gain) { m_gain = gain; }
