+++ src/blacksound/sampleprovider/volumesampleprovider.h
@@ -28,7 +28,7 @@ namespace BlackSound
             CVolumeSampleProvider(ISampleProvider *sourceProvider, QObject *parent = nullptr);
 
             //! \copydoc ISampleProvider::readSamples
-            virtual int readSamples(QVector<qint16> &samples, qint64 count) override;
+            virtual int readSamples(QVector<float> &samples, qint64 count) override;
 
             //! Volume @{
             double volume() const { return m_volume; }
